+++ src/common/config.cpp
@@ -187,9 +187,9 @@ bool wxConfigBase::Read(const wxString& key, float* val) const
     if ( !Read(key, &temp) )
         return false;
 
-    wxCHECK_MSG( fabs(temp) <= FLT_MAX, false,
+    wxCHECK_MSG( fabs(temp) <= double(FLT_MAX), false,
                      wxT("float overflow in wxConfig::Read") );
-    wxCHECK_MSG( (temp == 0.0) || (fabs(temp) >= FLT_MIN), false,
+    wxCHECK_MSG( temp == 0.0 || fabs(temp) >= double(FLT_MIN), false,
                      wxT("float underflow in wxConfig::Read") );
 
     *val = static_cast<float>(temp);
